👌 IMPROVE: re-factored useForm & formValidation into one module
