+++ src/Form.bs.js
@@ -1,129 +1,12 @@
 // Generated by BUCKLESCRIPT VERSION 5.0.6, PLEASE EDIT WITH CARE
 
 import * as $$Array from "bs-platform/lib/es6/array.js";
-import * as Curry from "bs-platform/lib/es6/curry.js";
 import * as React from "react";
 import * as Caml_chrome_debugger from "bs-platform/lib/es6/caml_chrome_debugger.js";
-import * as FormValidation$ReactHooksTemplate from "./FormValidation.bs.js";
+import * as UseForm$ReactHooksTemplate from "./UseForm.bs.js";
 
 function str(prim) {
   return prim;
-}
-
-var initialState = /* record */Caml_chrome_debugger.record([
-    "username",
-    "email",
-    "password"
-  ], [
-    "",
-    "",
-    ""
-  ]);
-
-function reducer(state, action) {
-  if (typeof action === "number") {
-    return initialState;
-  } else {
-    switch (action.tag | 0) {
-      case 0 : 
-          return /* record */Caml_chrome_debugger.record([
-                    "username",
-                    "email",
-                    "password"
-                  ], [
-                    action[0],
-                    state[/* email */1],
-                    state[/* password */2]
-                  ]);
-      case 1 : 
-          return /* record */Caml_chrome_debugger.record([
-                    "username",
-                    "email",
-                    "password"
-                  ], [
-                    state[/* username */0],
-                    action[0],
-                    state[/* password */2]
-                  ]);
-      case 2 : 
-          return /* record */Caml_chrome_debugger.record([
-                    "username",
-                    "email",
-                    "password"
-                  ], [
-                    state[/* username */0],
-                    state[/* email */1],
-                    action[0]
-                  ]);
-      
-    }
-  }
-}
-
-function useForm(formType, callback) {
-  var match = FormValidation$ReactHooksTemplate.useValidation(formType);
-  var allValid = match[2];
-  var validate = match[1];
-  var formRules = match[0];
-  var match$1 = React.useState((function () {
-          return false;
-        }));
-  var setIsSubmitting = match$1[1];
-  var isSubmitting = match$1[0];
-  var match$2 = React.useReducer(reducer, initialState);
-  var dispatch = match$2[1];
-  var state = match$2[0];
-  React.useEffect((function () {
-          var match = isSubmitting && allValid;
-          if (match) {
-            Curry._1(callback, /* () */0);
-            Curry._1(dispatch, /* ResetState */0);
-            return undefined;
-          } else {
-            Curry._1(setIsSubmitting, (function (param) {
-                    return false;
-                  }));
-            return undefined;
-          }
-        }), /* tuple */[
-        isSubmitting,
-        allValid
-      ]);
-  var handleChange = function (evt) {
-    evt.persist();
-    var match = evt.target.name;
-    switch (match) {
-      case "email" : 
-          Curry._1(dispatch, /* SetEmail */Caml_chrome_debugger.variant("SetEmail", 1, [evt.target.value]));
-          break;
-      case "password" : 
-          Curry._1(dispatch, /* SetPassword */Caml_chrome_debugger.variant("SetPassword", 2, [evt.target.value]));
-          break;
-      case "username" : 
-          Curry._1(dispatch, /* SetUsername */Caml_chrome_debugger.variant("SetUsername", 0, [evt.target.value]));
-          break;
-      default:
-        
-    }
-    console.log(formRules);
-    return /* () */0;
-  };
-  React.useEffect((function () {
-          Curry._1(validate, state);
-          return undefined;
-        }));
-  var handleSubmit = function (evt) {
-    evt.preventDefault();
-    return Curry._1(setIsSubmitting, (function (param) {
-                  return true;
-                }));
-  };
-  return /* tuple */[
-          state,
-          formRules,
-          handleChange,
-          handleSubmit
-        ];
 }
 
 function Form$FormErrors(Props) {
@@ -148,7 +31,7 @@ function Form(Props) {
     console.log("Form submitted");
     return /* () */0;
   };
-  var match = useForm(formType, logger);
+  var match = UseForm$ReactHooksTemplate.useForm(formType, logger);
   var handleChange = match[2];
   var formRules = match[1];
   var state = match[0];
@@ -216,9 +99,6 @@ var make = Form;
 
 export {
   str ,
-  initialState ,
-  reducer ,
-  useForm ,
   FormErrors ,
   make ,
